+++ R/monitor_loadAnnual.R
@@ -1,0 +1,103 @@
+#' @export
+#'
+#' @title Load annual monitoring data from all sources
+#'
+#' @param year Year [YYYY].
+#' @param archiveBaseUrl Base URL for monitoring v2 data files.
+#' @param archiveBaseDir Local base directory for monitoring v2 data files.
+#' @param QC_negativeValues Type of QC to apply to negative values.
+#'
+#' @return A \emph{mts_monitor} object with PM2.5 monitoring data.
+#'
+#' @description Combine annual data from AirNow, AIRSIS and WRCC:
+#'
+#' If \code{archiveDataDir} is defined, data will be loaded from this local
+#' archive. Otherwise, data will be loaded from the monitoring data repository
+#' maintained by the USFS AirFire team.
+#'
+#' Current year files loaded by this function are updated once per week.
+#'
+#' For the most recent data in the last 10 days, use \code{monitor_loadLatest()}.
+#'
+#' For daily updates covering the most recent 45 days, use \code{monitor_loadDaily()}.
+#'
+#' For data extended more than 45 days into the past, use \code{monitor_load()}.
+#'
+#' @note The AirNow data stream contains data may also be available from AIRSIS
+#' and WRCC. This can be detected by looking at the `locationID` associated with
+#' each time series. Wherever multiple time series share a `locationID`, the
+#' time series from AirNow or WRCC are removed so that each location is represented
+#' by a single time series coming from AirNow.
+#'
+# #' @seealso \code{\link{monitor_load}}
+#' @seealso \code{\link{monitor_loadDaily}}
+#' @seealso \code{\link{monitor_loadLatest}}
+#' @examples
+#' \dontrun{
+#' library(AirMonitor)
+#' # Fail gracefully if any resources are not available
+#' try({
+#'
+#' monitor_loadAnnual() %>%
+#'   monitor_filter(stateCode %in% CONUS) %>%
+#'   monitor_leaflet()
+#'
+#' }, silent = FALSE)
+#' }
+
+monitor_loadAnnual <- function(
+  year = NULL,
+  archiveBaseUrl = "http://data-monitoring_v2-c1.airfire.org/monitoring-v2",
+  archiveBaseDir = NULL,
+  QC_negativeValues = c("zero", "na", "ignore")
+) {
+
+  # ----- Validate parameters --------------------------------------------------
+
+  MazamaCoreUtils::stopIfNull(year)
+  parameterName = "PM2.5"
+
+  QC_negativeValues <- match.arg(QC_negativeValues)
+
+  if ( is.null(archiveBaseUrl) && is.null(archiveBaseDir) )
+    stop("one of 'archiveBaseUrl' or 'archiveBaseDir' must be defined")
+
+  # ----- Load data ------------------------------------------------------------
+
+  monitorList <- list()
+
+  try({
+    monitorList[["airnow"]] <- airnow_loadAnnual(year, parameterName, archiveBaseUrl, archiveBaseDir, QC_negativeValues)
+  }, silent = TRUE)
+
+  try({
+    monitorList[["airsis"]] <- airsis_loadAnnual(year, archiveBaseUrl, archiveBaseDir, QC_negativeValues)
+  }, silent = TRUE)
+
+  try({
+    monitorList[["wrcc"]] <- wrcc_loadAnnual(year, archiveBaseUrl, archiveBaseDir, QC_negativeValues)
+  }, silent = TRUE)
+
+  monitor_all <-
+    monitor_combine(monitorList) %>%
+    monitor_dropEmpty()
+
+  # ----- Remove duplicate locations -------------------------------------------
+
+  # NOTE:  Whenever we have multiple monitors reporting from the same location,
+  # NOTE:  we always favor the data fom AirNow over AIRSIS and WRCC.
+  # NOTE:  Because airnow comes first in monitor_combine() above, AirNow data
+  # NOTE:  will be preferentially retained.
+
+  ids <-
+    monitor_all$meta %>%
+    dplyr::distinct(.data$locationID, .keep_all = TRUE) %>%
+    dplyr::pull(.data$deviceDeploymentID)
+
+  monitor <- monitor_select(monitor_all, ids)
+
+  # ----- Return ---------------------------------------------------------------
+
+  return(monitor)
+
+}