+++ R/wrcc_loadDaily.R
@@ -2,11 +2,11 @@
 #'
 #' @title Load daily WRCC monitoring data
 #'
-#' @param archiveBaseUrl Base URL for annual EPA AQS data files.
-#' @param archiveBaseDir Local base directory for annual EPA AQS data files.
+#' @param archiveBaseUrl Base URL for monitoring v2 data files.
+#' @param archiveBaseDir Local base directory for monitoring v2 data files.
 #' @param QC_negativeValues Type of QC to apply to negative values.
 #'
-#' @return A \emph{ws_monitor} object with WRCC data.
+#' @return A \emph{mts_monitor} object with WRCC data.
 #'
 #' @description Loads pre-generated .rda files containing daily
 #' WRCC data.